feat 推荐接口完善

+++ browse/browse-server/src/main/java/org/ht/model/entity/VideoMetadata.java
@@ -86,7 +86,7 @@ public class VideoMetadata {
     /**
      * 视频时长（秒）
      */
-    private Integer durationSeconds;
+    private Long durationSeconds;
 
     /**
      * 播放次数
